# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Глафира Рекрутёр ATS",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/home/kpi": {
      "get": {
        "tags": [
          "home"
        ],
        "summary": "Get Kpi",
        "description": "Получить KPI для главной страницы",
        "operationId": "get_kpi_api_v1_home_kpi_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Период: week, month, quarter, year, all",
              "default": "month",
              "title": "Period"
            },
            "description": "Период: week, month, quarter, year, all"
          },
          {
            "name": "extended",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Включить расширенные метрики",
              "default": false,
              "title": "Extended"
            },
            "description": "Включить расширенные метрики"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeKpi"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/home/attention": {
      "get": {
        "tags": [
          "home"
        ],
        "summary": "Get Attention",
        "description": "Получить список вакансий, требующих внимания",
        "operationId": "get_attention_api_v1_home_attention_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttentionItem"
                  },
                  "type": "array",
                  "title": "Response Get Attention Api V1 Home Attention Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/home/events": {
      "get": {
        "tags": [
          "home"
        ],
        "summary": "Get Events",
        "description": "Получить ленту событий",
        "operationId": "get_events_api_v1_home_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Количество событий",
              "default": 30,
              "title": "Limit"
            },
            "description": "Количество событий"
          },
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по кандидату",
              "title": "Candidate Id"
            },
            "description": "Фильтр по кандидату"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Get Events Api V1 Home Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/home/pulse-summary": {
      "get": {
        "tags": [
          "home"
        ],
        "summary": "Get Pulse Summary",
        "description": "Получить сводку пульса",
        "operationId": "get_pulse_summary_api_v1_home_pulse_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Период: week, month, quarter, year, all",
              "default": "month",
              "title": "Period"
            },
            "description": "Период: week, month, quarter, year, all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PulseSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/home/sources": {
      "get": {
        "tags": [
          "home"
        ],
        "summary": "Get Sources",
        "description": "Получить топ источников кандидатов",
        "operationId": "get_sources_api_v1_home_sources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Период: week, month, quarter, year, all",
              "default": "month",
              "title": "Period"
            },
            "description": "Период: week, month, quarter, year, all"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceItem"
                  },
                  "title": "Response Get Sources Api V1 Home Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttentionHrItem": {
        "properties": {
          "employee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Employee Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "adapt_day": {
            "type": "integer",
            "title": "Adapt Day"
          },
          "risk_score": {
            "type": "integer",
            "title": "Risk Score"
          }
        },
        "type": "object",
        "required": [
          "employee_id",
          "full_name",
          "reason",
          "adapt_day",
          "risk_score"
        ],
        "title": "AttentionHrItem"
      },
      "AttentionItem": {
        "properties": {
          "vacancy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vacancy Id"
          },
          "vacancy_name": {
            "type": "string",
            "title": "Vacancy Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "vacancy_id",
          "vacancy_name",
          "kind",
          "text"
        ],
        "title": "AttentionItem"
      },
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "entities": {
            "items": {},
            "type": "array",
            "title": "Entities"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "text",
          "created_at"
        ],
        "title": "EventOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HomeKpi": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__home__KpiCard"
            },
            "type": "array",
            "title": "Cards"
          }
        },
        "type": "object",
        "required": [
          "period",
          "cards"
        ],
        "title": "HomeKpi"
      },
...
      "PulseSummary": {
        "properties": {
          "onboarding_count": {
            "type": "integer",
            "title": "Onboarding Count"
          },
          "onboarding_delta": {
            "type": "integer",
            "title": "Onboarding Delta"
          },
          "risk_split": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Risk Split"
          },
          "satisfaction_avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Satisfaction Avg"
          },
          "answered_pct": {
            "type": "number",
            "title": "Answered Pct"
          },
          "silent_pct": {
            "type": "number",
            "title": "Silent Pct"
          },
          "enps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enps"
          },
          "enps_delta": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enps Delta"
          },
          "attention_hr": {
            "items": {
              "$ref": "#/components/schemas/AttentionHrItem"
            },
            "type": "array",
            "title": "Attention Hr"
          }
        },
        "type": "object",
        "required": [
          "onboarding_count",
          "onboarding_delta",
          "risk_split",
          "answered_pct",
          "silent_pct",
          "attention_hr"
        ],
        "title": "PulseSummary"
      },
...
      "SourceItem": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "source",
          "count"
        ],
        "title": "SourceItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__schemas__home__KpiCard": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "delta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "delta_dir": {
            "type": "string",
            "title": "Delta Dir"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "key",
          "value",
          "delta_dir"
        ],
        "title": "KpiCard"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}